# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Signal AEO Admin Panel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ranking-reports:
    get:
      operationId: getRankingReports
      tags: [ranking-reports]
      summary: List ranking reports
      parameters:
        - name: clientId
          in: query
          schema:
            type: integer
        - name: keywordId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Ranking reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RankingReport"
    post:
      operationId: createRankingReport
      tags: [ranking-reports]
      summary: Create a ranking report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRankingReportBody"
      responses:
        "201":
          description: Ranking report created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RankingReport"
...
  /tasks:
    get:
      operationId: getTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [todo, in_progress, done]
        - name: priority
          in: query
          schema:
            type: string
            enum: [high, medium, low]
      responses:
        "200":
          description: Task list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /tasks/{id}/subtasks:
    post:
      operationId: createSubtask
      tags: [tasks]
      summary: Add a subtask
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubtaskBody"
      responses:
        "201":
          description: Subtask created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subtask"
  /tasks/{taskId}/subtasks/{subtaskId}:
    patch:
      operationId: updateSubtask
      tags: [tasks]
      summary: Update subtask done status
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
        - name: subtaskId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSubtaskBody"
      responses:
        "200":
          description: Updated subtask
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subtask"
...
components:
  schemas:
...
    RankingReport:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        keywordId:
          type: integer
        rankingPosition:
          type: integer
          nullable: true
        reasonRecommended:
          type: string
          nullable: true
        mapsPresence:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        clientName:
          type: string
          nullable: true
        keywordText:
          type: string
          nullable: true
        isInitialRanking:
          type: boolean
          nullable: true
      required: [id, clientId, keywordId, createdAt]
    CreateRankingReportBody:
      type: object
      properties:
        clientId:
          type: integer
        keywordId:
          type: integer
        rankingPosition:
          type: integer
          nullable: true
        reasonRecommended:
          type: string
          nullable: true
        mapsPresence:
          type: string
          nullable: true
        isInitialRanking:
          type: boolean
          nullable: true
      required: [clientId, keywordId]
...
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
          nullable: true
        status:
          type: string
          enum: [todo, in_progress, done]
        priority:
          type: string
          enum: [high, medium, low]
        notes:
          type: string
          nullable: true
        subtasks:
          type: array
          items:
            $ref: "#/components/schemas/Subtask"
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, title, status, priority, createdAt]
    CreateTaskBody:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
          nullable: true
        status:
          type: string
          enum: [todo, in_progress, done]
          default: todo
        priority:
          type: string
          enum: [high, medium, low]
          default: medium
        notes:
          type: string
          nullable: true
      required: [title]
    UpdateTaskBody:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
          nullable: true
        status:
          type: string
          enum: [todo, in_progress, done]
        priority:
          type: string
          enum: [high, medium, low]
        notes:
          type: string
          nullable: true
    Subtask:
      type: object
      properties:
        id:
          type: integer
        taskId:
          type: integer
        title:
          type: string
        done:
          type: boolean
      required: [id, taskId, title, done]
    CreateSubtaskBody:
      type: object
      properties:
        title:
          type: string
        done:
          type: boolean
          default: false
      required: [title]
    UpdateSubtaskBody:
      type: object
      properties:
        title:
          type: string
        done:
          type: boolean
...